Rename PlanSanityChecker to PlanChecker
